Chatbot Using JavaScript
This is a chatbot using javascript,html and css which is Just a simple  bot that you can chat with, written in only HTML , CSS and JavaScript.
I admit the code could've been written better and the algorithm isn't very advanced, but it's a good start for any beginners wanting to work on a conversational bot.